requirements.txt
torch
transformers
datasets
pillow
streamlit
